Generic params were swapped.

+++ price-calc-OOP/src/main/java/com/epam/brest/price/PriceCalculator.java
@@ -6,7 +6,7 @@ import com.epam.brest.requesters.RequesterAbstract;
 import java.util.List;
 import java.util.SortedMap;
 
-public class PriceCalculator extends RequesterAbstract<Long, Double> {
+public class PriceCalculator extends RequesterAbstract<Double, Long> {
 
     private final SortedMap<Double, Long> priceCurve;
